fixing multiple problems found during testing for #1
The XmlStore class was initially written to be used in a single-thread
context.  After adding the timed task queue, we had multiple threads
writing to BaseX, and this caused problems with the XmlStore class.

To solve this, AnalyzeTimedTaskWorker was altered to output its result
(the AST serialized in XML format). The result is then received by
the "gather" thread in AnalyzeTimedTaskQueue which writes it to BaseX.

The AbstractTimedTaskQueue now provides a callback method called
onTaskComplete which is run upon the succesful completion of a task
(it receives the result of the task as a parameter).
Some changes to the type parameters of AbstractTimedTaskQueue were also
made to allow the class user to specify their preferred result type for
the task.

The AnalyzeTimedTaskQueue now also has an implemented estimateTaskTimeout
method which computes an appropriate timeout depending on file size(larger
.php files are expected to take a longer time to parse).

From JDK7 onwards, it's recommended that threads check
if they're interrupted throughout their code by using
Thread.currentThread().isInterrupted() in order to take action if the
thread has been externally cancelled (we do this via Future.cancel()).
In this codebase, we use an auto-generated PHP parser (the ANTLR4 PHP
parser) and this parser does not check if the thread it's running in
was interrupted.
To solve this, the auto-generated PHPParser class was subclassed, and
the new class overrides the enterRule() method, and since this method
is called for every rule in the grammar, this offers a way to allow the
thread interruption to take place early throughout the execution.
The ParseCancellationException is thrown if the thread was interrupted,
allowing the thread to stop, in a way that aligns with the ANTLR API.